fix: Double tap start button bug removed
Previously, if the timer was running and the start
button was clicked again, another timer would run at
the same time. This caused the timer to look strange
(obviously as two timers were running but only one
was displayed). This change fixes the bug. Double tap
on the start button has "no" effect, the timer continues.